Update GitHub Workflow File
This is an attempt to port the Travis .yml contents to a GitHub Action.

It uses the install process for the latest Ren-C interpreter.  It could
also not do that and use some static version.  But it makes things more
useful for exercising the code to do it this way.

In order for it to work in terms of uploading to Amazon S3, there will
have to be information configured by logging into github.com and setting
the necessary "secrets" to get EIGHTH_AWS_ACCESS_KEY and
EIGHTH_AWS_SECRET_KEY.

But GitHub offers another feature, which is called "artifacts".  This is
not intended as long term storage but can be a way to get the results
out of a build from one job to another...or just let you download it
from the GitHub.com UI.

It would be interesting if this also used the ZIP facilities of Ren-C
instead of a separate tool, just as another test.