Eliminate loop for processing pdfseparate pages
Ghostscript does the separation correctly; the attempt to call
it was not working due to a typo (sDEVICE=eps2write was missing
a hyphen, e.g. -sDEVICE=eps2write)

With that adjustment it is not necessary to call pdfseparate and
loop over converting the individual EPS, and it can just delegate
to GhostScript.