Avoid double-zip on upload of images
Underlying GitHub's file uploading is that it uploads the
files as-is, and then a zip is generated on demand for each
artifact.

It is hence an interface limitation that if you only upload
one zip file, that the artifact will be doubly zipped.

People complain about this, but they don't seem to be jumping
on addressing it.  So this commit just gives in and lets the
artifacting do the zipping.