Switch .reb => .r, turn off syntax highlighting
The idea of switching Rebol extensions from .r to .reb did not really
have much of a point.  AI can tell the difference between Rebol and
R files well enough (if it wasn't already obvious), and they're not
likely to appear in the same project.  .r is less noisy

Syntax highlighting "keywords" in Rebol was always kind of silly, but
now the highlighting of Ren-C code is just broken since it has new
syntax entirely.  Turn it off.